# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "教务系统",
        "description": "教务系统接口文档",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login-in": {
            "post": {
                "tags": [
                    "用户认证"
                ],
                "summary": "用户登录",
                "operationId": "App\\Http\\Controllers\\Login\\LoginController::loginIn",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password",
                                    "type"
                                ],
                                "properties": {
                                    "username": {
                                        "description": "用户名",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "密码",
                                        "type": "string"
                                    },
                                    "type": {
                                        "description": "类型[teacher:教师,student:学生]",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "返回码",
                                            "type": "integer"
                                        },
                                        "message": {
                                            "description": "错误信息",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "properties": {
                                                "token": {
                                                    "description": "用户token",
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "description": "类型",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "description": "服务器响应的时间戳",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/register": {
            "post": {
                "tags": [
                    "用户认证"
                ],
                "summary": "教师注册",
                "operationId": "App\\Http\\Controllers\\Login\\LoginController::registerTeacher",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "用户名",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "邮箱",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "密码",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "返回码",
                                            "type": "integer"
                                        },
                                        "message": {
                                            "description": "错误信息",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "type": "object"
                                        },
                                        "extra": {
                                            "description": "返回额外数据",
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "description": "服务器响应的时间戳",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}